generate_nmc_cert: Use P256 curve by default.
In Go stdlib, RSA2048 is used by default.  RSA support was removed in our fork, but we neglected to set a default ECDSA curve, so the user had to choose a curve.  P256 is recommended by the Go devs and by us, so it seems to be a reasonable default.